+++ RELEASE_NOTES.md
@@ -9,6 +9,14 @@ KeyboardKit tries to honor the following rules when new versions are released:
 Breaking changes can still occur in minor versions, if the alternative is to not release new critical features or fixes.
 
 
+## 4.6.1
+
+### 🐛 Bug fixes
+
+* This version fixes a bug that caused the keyboard type to not change when typing into a `TextInputProxy` text field.   
+
+
+
 ## 4.6
 
 This version adds features that makes it possible to add text fields and text views to the keyboard extension and automatically redirect keyboard events there instead of using the hosting app.